+++ src/core/modes-utils.ts
@@ -26,19 +26,21 @@ export function joinLatex(segments: string[]): string {
     let sep = '';
     let result = '';
     for (const segment of segments) {
-        if (/[a-zA-Z*]/.test(segment[0])) {
-            // If the segment begins with a char that *could* be in a command
-            // name... insert a separator (if one was needed for the previous segment)
-            result += sep;
+        if (segment) {
+            if (/[a-zA-Z*]/.test(segment[0])) {
+                // If the segment begins with a char that *could* be in a command
+                // name... insert a separator (if one was needed for the previous segment)
+                result += sep;
+            }
+            // If the segment ends in a command...
+            if (/\\[a-zA-Z]+\*?$/.test(segment)) {
+                // ... potentially add a space before the next segment
+                sep = ' ';
+            } else {
+                sep = '';
+            }
+            result += segment;
         }
-        // If the segment ends in a command...
-        if (/\\[a-zA-Z]+\*?$/.test(segment)) {
-            // ... potentially add a space before the next segment
-            sep = ' ';
-        } else {
-            sep = '';
-        }
-        result += segment;
     }
     return result;
 }