+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## 0.56.0 (2020-08-22)
 
 ### New Features